+++ tests/timer_test.cc
@@ -6,8 +6,8 @@
 using std::string;
 
 static string s_log;
-static uint32_t s_now = 0;
-static uint32_t timer_get_time() {
+static uint64_t s_now = 0;
+static uint64_t timer_get_time() {
   return s_now;
 }
